Create Coding Exercise: Go to it!
Define a function findLine(prog, target) to perform the following. Assume prog is a list of strings containing a BASIC program, like the type generated by getBASIC(); assume target is a string containing a line number, which is the target of a GOTO statement. The function should return the index i (a number between 0 and len(prog)-1) such that prog[i] is the line whose label equals target. Hint 
Sample input/output: If you call
findLine(['10 GOTO 20','20 END'], '10')
then the output should be 0, since item 0 of the list is the line with label 10.