Create  Spy Decoder
Decoding
Once your friend gets the message, if they know the secret shift value S then it is easy for them to decipher the message: each letter is replaced by the one appearing S places earlier in the alphabet. For example, they would look at L, step back two positions in the alphabet and find J, which they now know is the first letter of your secret message. Again, they have to treat the alphabet as cyclic, assuming that Z comes before A