Create Primed for Takeoff
Write a program that defines the table isPrime we described above (notice that isPrime is a list, not a function). 
The grader will allow a longer-than-usual amount of time, 7 seconds, for your program to execute. However, simply using the isItPrime function will not be fast enough.
Hint -  https://en.wikipedia.org/wiki/Sieve_of_Eratosthenes